Divide Costs Fairly Among Participants

This function calculates the fair division of costs among participants based on their weights. It reads the participants' and costs' data from files, calculates the total costs for each participant or group, and returns the amounts to be paid after applying any payments that have already been made.

Usage

divide_fairly(
  participants_df = check_participants(),
  costs_df = check_costs(),
  pay_by = "group"
)

Arguments

pay_by

A string specifying whether to divide costs by "group" or "individual". Defaults to "group".

participants_file

Path to the CSV file containing participants' information. Defaults to a file within the package.

costs_file

Path to the CSV file containing costs' information. Defaults to a file within the package.

Value

A data frame containing the final payments for each person or group after minimizing the payments.

Details

The function starts by reading the participants and costs data. It calculates the total costs and distributes them based on the participants' weights. It also considers the payments already made. Depending on the pay_by argument, costs are divided either by group or by individual. The final step involves minimizing the payments using the minimize_payments function.

Examples

result <- divide_fairly(pay_by = "group")